fix wrong usage of styled-components
