+++ .babelrc
@@ -21,6 +21,9 @@
       "displayName": true,
       "preprocess": false
     }],
-    "babel-plugin-inline-import-graphql-ast"
+    "babel-plugin-inline-import-graphql-ast",
+    ["babel-plugin-root-import", {
+      "rootPathSuffix": "./"
+    }]
   ]
 }
