+++ components/AuthFields/styles.js
@@ -1,4 +1,5 @@
 import styled from 'styled-components';
+import * as T from '~/components/Theme';
 
 export const Main = styled.div`
   border-bottom: 1px solid #ececec;
@@ -15,6 +16,6 @@ export const Main = styled.div`
   }
 `;
 
-export const SubmitButton = styled.button`
+export const SubmitButton = T.Button.extend`
   opacity: ${({ touched }) => (touched ? 1 : 0.5)};
 `;
