+++ containers/CreatePost/styles.js
@@ -1,10 +1,12 @@
 import styled from 'styled-components';
+import { Button } from '~/components/Theme';
 
 // eslint-disable-next-line import/prefer-default-export
 export const Form = styled.form`
   border-bottom: 1px solid #ececec;
   padding-bottom: 20px;
   margin-bottom: 20px;
+
   > h1 {
     font-size: 20px;
   }
@@ -13,3 +15,4 @@ export const Form = styled.form`
     margin-bottom: 10px;
   }
 `;
+export { Button as SubmitButton };
