+++ containers/Header/styles.js
@@ -1,4 +1,4 @@
 import styled from 'styled-components';
 
 // eslint-disable-next-line import/prefer-default-export
-export const StyledHeader = styled.header`margin-bottom: 25px;`;
+export const Header = styled.header`margin-bottom: 25px;`;
