+++ package.json
@@ -98,15 +98,17 @@
   "devDependencies": {
     "babel-eslint": "7.2.3",
     "babel-plugin-inline-import-graphql-ast": "2.0.0",
+    "babel-plugin-root-import": "5.1.0",
     "cli-clear": "1.0.4",
     "enzyme": "2.9.1",
     "enzyme-to-json": "1.5.1",
     "eslint": "3.19.0",
     "eslint-config-airbnb": "15.1.0",
     "eslint-config-prettier": "2.3.0",
+    "eslint-import-resolver-babel-plugin-root-import": "https://github.com/bingqichen/eslint-import-resolver-babel-plugin-root-import",
     "eslint-loader": "1.9.0",
     "eslint-plugin-graphql": "1.3.0",
-    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-jest": "20.0.3",
     "eslint-plugin-jsx-a11y": "5.1.1",
     "eslint-plugin-prettier": "2.1.2",
