+++ pages/create.js
@@ -1,4 +1,4 @@
-import CreatePost from '../components/CreatePost';
+import CreatePost from '../containers/CreatePost';
 import withData from '../libraries/withData';
 import DefaultCon from '../containers/Default';
