+++ pages/index.js
@@ -1,4 +1,4 @@
-import PostList from '../components/PostList';
+import PostList from '../containers/PostList';
 import withData from '../libraries/withData';
 import DefaultCon from '../containers/Default';
